New behaviour: Disable prefixing by default
Prefixing now needs to be enabled on a
per-unit basis. This is the opposite
behaviour as before.

For now prefixing is enabled on the 7 SI
base units and 22 SI derived units.

The 7 base units are:

   second
   metre
   gram (SI uses kilogram, I use gram)
   ampere
   kelvin
   mole
   candela

The 22 derived units are:

  joule
  newton
  coulomb
  celsius
  katal
  lux
  watt
  pascal
  volt
  weber
  tesla
  lumen
  hertz
  becquerel
  gray
  ohm
  henry
  farad
  steradian
  sievert
  siemens
  radian

Prefix scaling is also enabled for other units
such as bits, bytes, litres, etc...

Move factor to 2nd argument.

Update README